Server_Review/requirements.txt
grpcio
googleapis-common-protos
flask
redis
numpy
